requirements.txt
argon2-cffi==20.1.0
cffi==1.14.5
psutil==5.8.0
pycparser==2.20
pycryptodome==3.10.1
six==1.15.0
